requirements.txt
# Core Dependencies
folium==0.14.0
geopy==2.3.0
numpy==1.24.3
requests==2.31.0
PyYAML==6.0
cachetools==5.3.1

# Development Extras
jupyter==1.0.0
ipython==8.14.0
matplotlib==3.7.2
pytest==7.4.0

# Optional (for advanced features)
scipy==1.10.1       # For more accurate coverage calculations
pandas==2.0.3       # For data analysis
flask==2.3.2        # For web interface
python-dotenv==1.0.0 # For environment variables